Adds set -e bash option and Puts catalogd_prep in a subshell with checks on cd
In a subshell we make sure that we don't have to worry about
explicitly jumping out of the current filesystem diectory in case of
an error.

Put a mkdir -p there so that it is always works.
There is a valid debate on if we want to make sure the script doesn't
work more than once for the creation of the 4{STAGING_DIR} Hmmm
:thinking_face: