Modified makefile patching using _here string_ & wrote a "main" function
Using a here string reduces having to manage the makefile patch file
on the filesystem - one less thing to worry about. Putting the
argument values in one time as variables also makes things more DRY,
as you specify them in one place at the top of the function.

I should have made the following a diff commit, but didn't know quite
how to work the chunking right.

Created a main function to wrap up all the "loose" logic at the bottom
of the script. This is not entirely necessary but helps with
readability, IMHO.