:golf: Reading list of files to move via ls -a with while read loop
Wanted to remove the `FILES` variable holding the output of `ls -a`.
By using `while read` over a process substitution stream (a named
FIFO pipe) from the `ls -a` command we are able to directly loop over
the output and process it.

A bit of golfing :golf: regarding the long if conditional test by
reducing that to a regex set test (note: for bash version 3 or
greater), we can make it a bit cleaner.  Hmmm... at least fewer
keystrokes / shorter line. :upside_down_face:

Made the `if` conditional into affirmative test, such the branch we
use the most is in the affirmative and what we do the least in the
negative.  Ultimately, as the TODO states, we should remove that in
the final version - at least put it behind DEUBG or VERBOSE flag so
the output is not as spammy.