:hammer_and_wrench: Added undo functionality to accelerate dev cycle
To facilitate a faster dev cycle (as well as it would be handy) there
is now "undo" functionality. Added some arg handling to be able to
trigger this functionality and while we are at it, throw in a usage
that recapitulates the README with some standard unix-like arg
explanation.

:warn:
Do not use the --undo if you have pushed for goodness sake

:memo:
At this point the scrip it is NOT performing perfectly.

* sed is complaining
* build is net yet successful (at least for me)