:hammer_and_wrench: GNU Sed is reqired along with other tools
On am mac sed is not necessarily gnu sed.  The sed version on the mac
does not work as expected in the script, therefore we do some checking
on the operating system as well as what tools are needed in the script
and we check for them and set them appropriately.
In particular we check if you are on a mac and if you have gsed if so
we set the ${sed_cmd} to gsed.  Otherwise you use the sed you
have (ex when you are on a *NIX box).

There are other odds and ends with output lines and having a VERBOSE
and DEBUG environment flags that we can use.