:butterfly: User Experience with prerequisite tools
To offer a better user experience...

The idea here is that the user should be notified up front about all
the tools that they need to have.  Specifically, all the tools that
they don't have should be shared with the user at one time.  As
opposed to the scenario where you have 4 tools missing, but you have
to run the script 4x and each time it tells you what the next missing
tool is. So though it is usually advantageous to set the shell option
for `errexit` to be on (the script exits entirely if any command
errors out) - this time we want to tolerate multiple errors until we
cascade through the entire set of checks and output all the
information. Subsequent to that, then we set the `errexit` option.

Yes, I also made the output here a bit fancy with bold yellow fonts
for the missing tool, :nail_care: indeed.