⛳ A bit of golfing and some opinionated best practices updates 😏  (#2)
* (Best practice) Use /usr/bin/env  bash

* :nail_care: (Best practices) no need to use the word "function"

It is unnecessary to use the world "function" it is redundant with
"()".

* :nail_care: Cosmetic - took out spaces between commands

IMHO makes the code more readable.

* Adds set -e bash option and Puts catalogd_prep in a subshell with checks on cd

In a subshell we make sure that we don't have to worry about
explicitly jumping out of the current filesystem diectory in case of
an error.

Put a mkdir -p there so that it is always works.
There is a valid debate on if we want to make sure the script doesn't
work more than once for the creation of the 4{STAGING_DIR} Hmmm
:thinking_face:

* Good practice to put the command to be tested directly in the if statement

Instead of checking for the return value and using that in an if
statement, call the command directly in the if statement.

* pipline the function calls

We want to avoid partial failures by stopping progress if there is
one.

* Modified makefile patching using _here string_ & wrote a "main" function

Using a here string reduces having to manage the makefile patch file
on the filesystem - one less thing to worry about. Putting the
argument values in one time as variables also makes things more DRY,
as you specify them in one place at the top of the function.

I should have made the following a diff commit, but didn't know quite
how to work the chunking right.

Created a main function to wrap up all the "loose" logic at the bottom
of the script. This is not entirely necessary but helps with
readability, IMHO.

* :bug: quick fix s/elif/else

* :golf: Reading list of files to move via ls -a with while read loop

Wanted to remove the `FILES` variable holding the output of `ls -a`.
By using `while read` over a process substitution stream (a named
FIFO pipe) from the `ls -a` command we are able to directly loop over
the output and process it.

A bit of golfing :golf: regarding the long if conditional test by
reducing that to a regex set test (note: for bash version 3 or
greater), we can make it a bit cleaner.  Hmmm... at least fewer
keystrokes / shorter line. :upside_down_face:

Made the `if` conditional into affirmative test, such the branch we
use the most is in the affirmative and what we do the least in the
negative.  Ultimately, as the TODO states, we should remove that in
the final version - at least put it behind DEUBG or VERBOSE flag so
the output is not as spammy.

* :bow: Sets bash flags - classic euo pipefail

* :hammer_and_wrench: Added undo functionality to accelerate dev cycle

To facilitate a faster dev cycle (as well as it would be handy) there
is now "undo" functionality. Added some arg handling to be able to
trigger this functionality and while we are at it, throw in a usage
that recapitulates the README with some standard unix-like arg
explanation.

:warn:
Do not use the --undo if you have pushed for goodness sake

:memo:
At this point the scrip it is NOT performing perfectly.

* sed is complaining
* build is net yet successful (at least for me)

* :hammer_and_wrench: GNU Sed is reqired along with other tools

On am mac sed is not necessarily gnu sed.  The sed version on the mac
does not work as expected in the script, therefore we do some checking
on the operating system as well as what tools are needed in the script
and we check for them and set them appropriately.
In particular we check if you are on a mac and if you have gsed if so
we set the ${sed_cmd} to gsed.  Otherwise you use the sed you
have (ex when you are on a *NIX box).

There are other odds and ends with output lines and having a VERBOSE
and DEBUG environment flags that we can use.

* :butterfly: User Experience with prerequisite tools

To offer a better user experience...

The idea here is that the user should be notified up front about all
the tools that they need to have.  Specifically, all the tools that
they don't have should be shared with the user at one time.  As
opposed to the scenario where you have 4 tools missing, but you have
to run the script 4x and each time it tells you what the next missing
tool is. So though it is usually advantageous to set the shell option
for `errexit` to be on (the script exits entirely if any command
errors out) - this time we want to tolerate multiple errors until we
cascade through the entire set of checks and output all the
information. Subsequent to that, then we set the `errexit` option.

Yes, I also made the output here a bit fancy with bold yellow fonts
for the missing tool, :nail_care: indeed.

* :nail_care: Added info in help about env vars that can be set

* :nail_care: Cosmetic - indentation / formatting

* Able to anchor catalogd and operator-controller locations explicitly - updated README

closes OchiengEd/olm-monorepo-migration#3

* :fixed: Fixed typo s/catlogd/catalogd/g